Last changes, changed names of tests and added new allure report
